+++ src/main/java/com/meiye/model/role/AuthPermission.java
@@ -23,14 +23,14 @@ public class AuthPermission extends ParentEntity {
 	private String levelId;
 	private String name;
 	private String code;
-	private Byte type;
+	private Integer type;
 	private String url;
 	private Integer sort;
-	private Byte platform;
-	private Byte supportVersion;
-	private Byte sourceFlag;
-	private Byte groupFlag;
-	private Byte checked;
+	private Integer platform;
+	private Integer supportVersion;
+	private Integer sourceFlag;
+	private Integer groupFlag;
+	private Integer checked;
 	private String zoneCode;
 
 }
